Inline `host` module into `model`
I'm going to need this name for something else.
