+++ crates/fj-host/src/model.rs
@@ -7,7 +7,7 @@ use std::{
 
 use fj::{abi, version::RawVersion};
 
-use crate::{host::Host, platform::HostPlatform, Parameters};
+use crate::{platform::HostPlatform, Parameters};
 
 /// Represents a Fornjot model
 pub struct Model {
@@ -152,6 +152,36 @@ pub struct Evaluation {
 
     /// The time it took to compile the shape, from the Cargo output
     pub compile_time: String,
+}
+
+pub struct Host<'a> {
+    args: &'a Parameters,
+    model: Option<Box<dyn fj::models::Model>>,
+}
+
+impl<'a> Host<'a> {
+    pub fn new(parameters: &'a Parameters) -> Self {
+        Self {
+            args: parameters,
+            model: None,
+        }
+    }
+
+    pub fn take_model(&mut self) -> Option<Box<dyn fj::models::Model>> {
+        self.model.take()
+    }
+}
+
+impl<'a> fj::models::Host for Host<'a> {
+    fn register_boxed_model(&mut self, model: Box<dyn fj::models::Model>) {
+        self.model = Some(model);
+    }
+}
+
+impl<'a> fj::models::Context for Host<'a> {
+    fn get_argument(&self, name: &str) -> Option<&str> {
+        self.args.get(name).map(|s| s.as_str())
+    }
 }
 
 fn package_associated_with_directory<'m>(
